requirements.txt
torch
torchvision
transformers
sentence-transformers
datasets
scikit-learn
numpy
pillow
tqdm
matplotlib
seaborn
requests
